+++ smarts/core/sensors.py
@@ -575,9 +575,20 @@ class Sensors:
             if on_lane.road in route_roads:
                 return (False, is_wrong_way)
 
-        for lanes in nearest_lane.incoming_lanes:
-            if on_lane in lanes:
-                return (False, is_wrong_way)
+        # Check for case if there was an early merge into another incoming lane. This means the
+        # vehicle should still be following the lane direction.
+        if not is_wrong_way:
+            for lanes in nearest_lane.outgoing_lanes:
+                for lane in lanes:
+                    if lane.in_junction:
+                        for out_lane in lane.outgoing_lanes:
+                            if out_lane in route_roads:
+                                break
+                        else:
+                            continue
+                    if not lane in route_roads:
+                        continue
+                    return (True, is_wrong_way)
 
         # Vehicle is completely off-route
         return (True, is_wrong_way)